requirements.txt
sqlalchemy
asyncpg
python-dotenv
httpx
pydantic
apscheduler
fastapi
uvicorn
tenacity